+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable changes to Flightdeck for 11ty will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 2023-12-27
+
+### Adds
+- `{% Image 'src' 'alt text' 'sizes: optional' %}` shortcode for responsive images
+    - shortcode uses the eleventy-img plugin to generate responsive images
+    - see `_flightdeck/shortcodes/image.js` for usage
+- Filters update
+    - adds stripFileExtension.js
+    - adds postLimit.js
+    - removes limit.js
+- Adds `_flightdeck/transforms/images.js` for image processing that live inside of `./src/assets/images`
+    - uses the eleventy-img plugin to optimize images
+    - see `_flightdeck/transforms/images.js` for usage
+
+### Changed
+- `_flightdeck/transforms/images.js` now uses the eleventy-img plugin for image optimization
+- `_flightdeck/filters/limit.js` is removed in favor of `postLimit.js` and `stripFileExtension.js`
+
+### Chore
+- 🧹 Package updates:
+    - removes unused packages
+    - Updates esbuild to v0.19.10
+
 ## [0.4.2] - 2023-12-22
 
 Attempting to improve performance of image processing.